+++ mycodo/controller_pid.py
@@ -638,7 +638,9 @@ class PIDController(threading.Thread):
 
                 if self.control_variable > 0:
                     # Determine if the output should be PWM or a duration
-                    if self.raise_output_type in ['pwm', 'command_pwm']:
+                    if self.raise_output_type in ['pwm',
+                                                  'command_pwm',
+                                                  'python_pwm']:
                         self.raise_duty_cycle = float("{0:.1f}".format(
                             self.control_var_to_duty_cycle(self.control_variable)))
 
@@ -667,6 +669,7 @@ class PIDController(threading.Thread):
                             self.control_var_to_duty_cycle(self.control_variable))
 
                     elif self.raise_output_type in ['command',
+                                                    'python',
                                                     'wired',
                                                     'wireless_rpi_rf']:
                         # Ensure the output on duration doesn't exceed the set maximum
@@ -695,7 +698,9 @@ class PIDController(threading.Thread):
                             self.control_variable)
 
                 else:
-                    if self.raise_output_type in ['pwm', 'command_pwm']:
+                    if self.raise_output_type in ['pwm',
+                                                  'command_pwm',
+                                                  'python_pwm']:
                         self.control.output_on(self.raise_output_id,
                                                duty_cycle=0)
 
@@ -707,7 +712,9 @@ class PIDController(threading.Thread):
 
                 if self.control_variable < 0:
                     # Determine if the output should be PWM or a duration
-                    if self.lower_output_type in ['pwm', 'command_pwm']:
+                    if self.lower_output_type in ['pwm',
+                                                  'command_pwm',
+                                                  'python_pwm']:
                         self.lower_duty_cycle = float("{0:.1f}".format(
                             self.control_var_to_duty_cycle(abs(self.control_variable))))
 
@@ -744,6 +751,7 @@ class PIDController(threading.Thread):
                             stored_control_variable)
 
                     elif self.lower_output_type in ['command',
+                                                    'python',
                                                     'wired',
                                                     'wireless_rpi_rf']:
                         # Ensure the output on duration doesn't exceed the set maximum
@@ -779,7 +787,9 @@ class PIDController(threading.Thread):
                             stored_control_variable)
 
                 else:
-                    if self.lower_output_type in ['pwm', 'command_pwm']:
+                    if self.lower_output_type in ['pwm',
+                                                  'command_pwm',
+                                                  'python_pwm']:
                         self.control.output_on(self.lower_output_id,
                                                duty_cycle=0)